Remove redundant, tidy
